+++ test/conftest.py
@@ -4,8 +4,8 @@ import geopandas as gpd
 import pytest
 import responses
 import shapely
-from climatoology.base.computation import ComputationScope
 from climatoology.base.baseoperator import AoiProperties
+from climatoology.base.computation import ComputationScope
 
 from bikeability.indicators.path_categories import PathCategory
 from bikeability.input import ComputeInputBikeability
@@ -63,9 +63,12 @@ def ohsome_api(responses_mock):
     with (
         open('resources/test/ohsome_line_response.geojson', 'r') as line_file,
         open('resources/test/ohsome_polygon_response.geojson', 'r') as polygon_file,
+        open('resources/test/ohsome_node_response.geojson', 'r') as node_file,
     ):
         line_body = line_file.read()
         polygon_body = polygon_file.read()
+        node_body = node_file.read()
+
     responses_mock.post(
         'https://api.ohsome.org/v1/elements/geometry',
         body=line_body,
@@ -76,6 +79,12 @@ def ohsome_api(responses_mock):
         body=polygon_body,
         match=[filter_start_matcher('geometry:polygon')],
     )
+    responses_mock.post(
+        'https://api.ohsome.org/v1/elements/geometry',
+        body=node_body,
+        match=[filter_start_matcher('geometry:point')],
+    )
+
     return responses_mock